Tidy up and use TLS instead of static mut
