+++ src/user_crate/state_loaded.rs
@@ -1,42 +1,34 @@
-use crate::user_crate::CrateState;
-use libloading::{Library, Symbol};
+use crate::{user_crate::CrateState};
+use libloading::os::unix::{Library, Symbol};
 use pgx::pg_sys;
 use std::path::Path;
 
-impl<'a> CrateState for StateLoaded<'a> {}
+impl CrateState for StateLoaded {}
 
 #[must_use]
-pub struct StateLoaded<'a> {
+pub struct StateLoaded {
+    #[allow(dead_code)] // Mostly for debugging
     fn_oid: pg_sys::Oid,
+    #[allow(dead_code)] // We must hold this handle for `symbol`
     library: Library,
-    symbol: once_cell::sync::OnceCell<
-        Symbol<'a, unsafe extern "C" fn(pg_sys::FunctionCallInfo) -> pg_sys::Datum>,
-    >,
+    symbol: Symbol<unsafe extern "C" fn(pg_sys::FunctionCallInfo) -> pg_sys::Datum>,
 }
 
-impl<'a> StateLoaded<'a> {
+impl StateLoaded {
     #[tracing::instrument(level = "debug", skip_all)]
     pub(crate) unsafe fn load(fn_oid: pg_sys::Oid, shared_object: &Path) -> eyre::Result<Self> {
         let library = Library::new(shared_object)?;
+        let symbol_name = crate::plrust::symbol_name(fn_oid);
+        let symbol = library.get(symbol_name.as_bytes())?;
 
-        let this = Self {
+        Ok(Self {
             fn_oid,
             library,
-            symbol: Default::default(),
-        };
-
-        Ok(this)
+            symbol,
+        })
     }
 
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub unsafe fn symbol(
-        &'a self,
-    ) -> eyre::Result<&'a Symbol<'a, unsafe extern "C" fn(pg_sys::FunctionCallInfo) -> pg_sys::Datum>>
-    {
-        self.symbol.get_or_try_init(|| {
-            let symbol_name = crate::plrust::symbol_name(self.fn_oid);
-            let symbol = self.library.get(symbol_name.as_bytes())?;
-            Ok(symbol)
-        })
+    pub unsafe fn evaluate(&self, fcinfo: pg_sys::FunctionCallInfo) -> pg_sys::Datum {
+        (self.symbol)(fcinfo)
     }
 }
